+++ flask_restplus/namespace.py
@@ -2,7 +2,15 @@
 from __future__ import unicode_literals
 
 
-class ApiNamespace(object):
+class Namespace(object):
+    '''
+    Group resources together
+
+    :param Api api: The API on which the namespace is attached
+    :param str name: The namespace name
+    :param str description: An optionnale short description
+    :param str path: An optionnal prefix path. If not provided, prefix is ``/+name``
+    '''
     def __init__(self, api, name, description=None, endpoint=None, path=None, **kwargs):
         self.api = api
         self.name = name